General code refactor
Fix metadata, add update notes and check for updates. Renamed file for future update's sake. Added 'Install Now' button to README.
