+++ README.md
@@ -1,8 +1,12 @@
+# <p align="center">WME MapCommentGeometry</p>
+
 This script creates a map comment around a single selected road segment.
 
-Instructions
-1) install this script in TamperMonkey
-2) select a road in WME (you can create a new one in the shape of the map comment you want to make)
-3) click the "Make Map Comment" button at the bottom of the left pane
-4) create a new Map Comment area or select an existing one
-5) click the "Map Comment on Road" button on the left pane
+## Instructions
+1) Install this script in Tampermonkey.
+2) Select a road in WME (you can create a new one in the shape of the map comment you want to make).
+3) Click the "Make Map Comment" button at the bottom of the left pane.
+4) Create a new Map Comment area or select an existing one.
+5) Click the "Map Comment on Road" button on the left pane.
+
+<p align="center"><a href="https://raw.githubusercontent.com/YULWaze/WME-MapCommentGeometry/main/WME%20MapCommentGeometry.user.js"><img src="https://i.ibb.co/JzHFKzj/button-install-here.png"></a></p>
